uart/hard_uart: Enable UDR Ready Interrupt after incrementing tx_buf.tail
Looks like if we do it before then characters will be lost/delayed.

Examples:

	while(1){
		Serial.send('A');
		Serial.send('b');
		_delay_ms(1000);
	}

The behaviour will be in the first loop, only 'A' is transmitted. In the second loop, 'b' from the first loop and 'Ab' from the second loop is transmitted

Inspiration for this fix taken from andygock/avr-uart in uart.c